+++ CHANGELOG.md
@@ -5,6 +5,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][Unreleased]
 
+## [0.68.0][0.68.0] - 2026-03-09
+
+Added:
+
+1. Support Telegram Bot API v9.3 (by @danielperez9430)
+   * `#getUserGifts`
+   * `#getChatGifts`
+   * `#sendMessageDraft` (by @xjx0106)
+   * `#repostStory`
+
+2. Fixed method:
+   * unpinAllGeneralForumTopicMessages
+
 ## [0.67.0][0.67.0] - 2025-12-13
 
 Added:
@@ -578,4 +591,5 @@ Fixed:
 [0.65.1]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v0.65.1
 [0.66.0]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v0.66.0
 [0.67.0]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v0.67.0
-[Unreleased]:https://github.com/yagop/node-telegram-bot-api/compare/v0.67.0...master
+[0.68.0]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v0.68.0
+[Unreleased]:https://github.com/yagop/node-telegram-bot-api/compare/v0.68.0...master